+++ src/config.php
@@ -36,6 +36,9 @@ return [
     // Should the query string be stripped from the saved statistics source URLs?
     'stripQueryStringFromStats'   => true,
 
+    // Should the anonymous ip address of the client causing a 404 be recorded?
+    'recordRemoteIp' => true,
+
     // How many stats should be stored
     'statsStoredLimit' => 1000,